+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Thu, 27 Feb 2025 01:12:41 GMT. */
+/* Web components. File generated at Thu, 27 Feb 2025 17:35:40 GMT. */
 export * from './collapsible-content-panel/collapsible-content-panel.component';
 export * from './fixed-scroll-panel/fixed-scroll-panel.component';
 export * from './command-prompt/command-prompt.component';