+++ src/EvaluatorConfiguration.ts
@@ -1,3 +1,4 @@
+import type * as PlotlyType from 'plotly.js';
 import './fetchPolyfill';
 import './showOpenFilePickerPolyfill';
 import openFileDialog from './openFileDialog';
@@ -265,7 +266,7 @@ export const _plotData: Plotly.Data[] = [];
 
 export const outputFunction: { [k: string]: Function } = {
     plot: function (parent: string): void {
-        DynamicModule.use('plotly', (Plotly: any) => {
+        DynamicModule.use('plotly', async (Plotly: Promise<typeof PlotlyType>) => {
             const trace1 = {
                 x: [1, 2.5, 3, 4],
                 y: [10, 15, 13, 17],
@@ -278,13 +279,13 @@ export const outputFunction: { [k: string]: Function } = {
                 type: 'scatter',
             };
 
-            const data = [trace1, trace2];
-            Plotly.newPlot(parent, data);
+            const data = [trace1, trace2] as Plotly.Data[];
+            (await Plotly).newPlot(parent, data);
         });
         insertOutput.type = '';
     },
     plot3: function (parent: string): void {
-        DynamicModule.use('plotly', (Plotly: any) => {
+        DynamicModule.use('plotly', async (Plotly: Promise<typeof PlotlyType>) => {
             const line3d = {
                 x: [] as number[],
                 y: [] as number[],
@@ -307,18 +308,18 @@ export const outputFunction: { [k: string]: Function } = {
                 line3d.z.push(z);
             }
 
-            const data = [line3d];
+            const data = [line3d] as Plotly.Data[];
 
             const layout = {
                 autosize: true,
             };
-            Plotly.newPlot(parent, data, layout);
+            (await Plotly).newPlot(parent, data, layout);
         });
         insertOutput.type = '';
     },
     surf: function (parent: string): void {
         insertOutput.type = '';
-        DynamicModule.use('plotly', (Plotly: any) => {
+        DynamicModule.use('plotly', async (Plotly: Promise<typeof PlotlyType>) => {
             function custom(x: number, y: number) {
                 return Math.sin(x / 50) * Math.cos(y / 50) * 50 + 50;
             }
@@ -345,25 +346,25 @@ export const outputFunction: { [k: string]: Function } = {
                 }
             }
 
-            const data = [surface1];
+            const data = [surface1] as Plotly.Data[];
 
             const layout = {
                 title: '3D Plot',
                 autosize: true,
             };
-            Plotly.newPlot(parent, data, layout);
+            (await Plotly).newPlot(parent, data, layout);
         });
         insertOutput.type = '';
     },
     plot2d: function (output: HTMLElement): void {
-        DynamicModule.use('plotly', (Plotly: any) => {
+        DynamicModule.use('plotly', async (Plotly: Promise<typeof PlotlyType>) => {
             const trace = {
                 x: plotData.X,
                 y: plotData.data,
                 type: 'lines',
             };
 
-            const data = [trace];
+            const data = [trace] as Plotly.Data[];
             const layout = {};
             const config = {
                 displayModeBar: false, // Mostra ou esconde a barra
@@ -371,21 +372,21 @@ export const outputFunction: { [k: string]: Function } = {
                 staticPlot: false, // Se `true`, o gráfico fica estático, sem interações
                 // scrollZoom: true, // Permite zoom com a roda do mouse
             };
-            Plotly.newPlot(output, data, layout, config);
+            (await Plotly).newPlot(output, data, layout, config);
         });
         insertOutput.type = '';
     },
     histogram: function (parent: string): void {
-        DynamicModule.use('plotly', (Plotly: any) => {
+        DynamicModule.use('plotly', async (Plotly: Promise<typeof PlotlyType>) => {
             const histogram = {
                 x: plotData.X,
                 y: plotData.data,
                 type: 'bar',
             };
 
-            const data = [histogram];
+            const data = [histogram] as Plotly.Data[];
 
-            Plotly.newPlot(parent, data);
+            (await Plotly).newPlot(parent, data);
         });
         insertOutput.type = '';
     },